requirements.txt
boto==2.48.0
boto3==1.9.6
botocore==1.12.72
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Django==1.11.10
django-admin-honeypot==1.1.0
django-allauth==0.38.0
django-bootstrap3==11.0.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-extensions==2.1.3
django-misaka==0.2.1
django-storages==1.7.1
django-widget-tweaks==1.4.2
docutils==0.14
futures==3.2.0
gunicorn==19.9.0
houdini.py==0.1.0
idna==2.8
jmespath==0.9.3
misaka==2.1.1
oauthlib==2.1.0
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
Pygments==2.3.1
python-dateutil==2.7.5
python-decouple==3.1
python-openid==2.2.5
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.12.0
typing==3.6.6
unicode-slugify==0.1.3
Unidecode==1.0.23
urllib3==1.24.1
